+++ project.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 clear
 pip install -e .
+venv/bin/pip install prefact --upgrade
 venv/bin/pip install vallm --upgrade
 venv/bin/pip install redup --upgrade
 venv/bin/pip install glon --upgrade
@@ -20,3 +21,4 @@ venv/bin/redup scan . --format toon --output ./project
 #venv/bin/vallm batch ./src --recursive --semantic --model qwen2.5-coder:7b
 #vallm batch --parallel .
 venv/bin/vallm batch . --recursive --format toon --output ./project
+venv/bin/prefact -a